mod:レッスンの状態変更機能の追加をいたしました。
