+++ features/lessonAttendance/hooks/useUpdateLessonAttendance.ts
@@ -3,14 +3,14 @@ import { Lesson } from "@/features/lesson/types/Lesson";
 import { LessonAttendance } from "@/features/lesson/types/LessonAttendance";
 
 type Args = {
-    lesson_attendance_id: number | null | undefined;
-    status: string | null | undefined;
+    lesson_attendance_id: number | undefined;
+    status: string;
 }
 
 export const useUpdateLessonAttendance = ({ lesson_attendance_id, status }: Args) => {
     console.log(lesson_attendance_id);
     console.log(status);
-    Axios.patch('api/proxy/api/v1/lesson-attendance', {
+    Axios.patch('api/proxy/api/v1/lesson_attendance', {
         lesson_attendance_id: lesson_attendance_id,
         status: status
     }).then((res) => {
