+++ hooks/useFetchCourse.ts
@@ -1,8 +1,8 @@
 import { Axios } from '@/lib/api';
 import { Chapter } from '@/types/Chapter';
 import { Course } from '@/features/course/types/Course';
-import { Lesson } from '@/types/Lesson';
-import { LessonAttendance } from '@/types/LessonAttendance';
+import { Lesson } from '@/features/lesson/types/Lesson';
+import { LessonAttendance } from '@/features/lessonAttendance/types/LessonAttendance';
 import { useEffect, useState } from 'react';
 
 type Data = Course & {
