--- types/Lesson.ts
@@ -1,6 +1,0 @@
-export type Lesson = {
-  lesson_id: number;
-  remarks: string;
-  title: string;
-  url: string;
-};
